[FEAT/#28] 코드 정리
