+++ BiRTH/View/FriendListView/FriendListView.swift
@@ -1,68 +1,64 @@
 //
-//  FriendListView.swift
+//  Temp_FriendListView.swift
 //  BiRTH
 //
-//  Created by Cho YeonJi on 11/21/24.
+//  Created by 이소현 on 11/28/24.
 //
 
 import SwiftUI
 
 struct FriendListView: View {
-    @State private var text = ""
-    @State var viewMode: ListModeToggle.ViewMode = .grid
+    @Environment(\.managedObjectContext) var viewContext
+    
+    @State var text: String = ""
+    @State var selectedViewMode: ViewMode = .grid
+    @State var isGridView = ViewMode.noFriend
     @State var sortingMethod = "생일 가까운 순"
-    @State var isGridView = true
-    @State var isDdaySort = true
-    @State private var relationshipTag = [""]
-
-    @FetchRequest(entity: BFriend.entity(), sortDescriptors: [NSSortDescriptor(keyPath: \BFriend.birth, ascending: true)])
-    private var bFriend: FetchedResults<BFriend>
-
-    @State private var isshowingSheetForCreatingTag = false
-    @FetchRequest(entity: BTag.entity(), sortDescriptors: [])
-    private var bTags: FetchedResults<BTag>
+    @State var tagName = [""]
+    @State var tagColor = [""]
     
-    let ddayUtils = DdayUtils()
+    @FetchRequest(entity: BTag.entity(), sortDescriptors: []) var bTags: FetchedResults<BTag>
+    @FetchRequest(entity: BFriend.entity(), sortDescriptors: []) var bFriend: FetchedResults<BFriend>
     
+    let gridShapes: [any Shape] = [CustomRectangle1(), CustomRectangle2()]
     
-    var sortedFriends: [BFriend] {
-        if isDdaySort {
-            return bFriend.sorted { friend1, friend2 in
-                guard let birth1 = friend1.birth, let birth2 = friend2.birth else { return false }
-                let dday1 = ddayUtils.calculateDday(birth: birth1)
-                let dday2 = ddayUtils.calculateDday(birth: birth2)
-                return dday1 < dday2
-            }
-        } else {
-            return bFriend.sorted { ($0.name ?? "") < ($1.name ?? "") }
-        }
-    }
-    
-
-
     var body: some View {
         NavigationStack {
             VStack {
-                SearchBarForFriendListView(text: $text)
-                HeaderForFriendListView(viewMode: $viewMode, isGridView: $isGridView, isDdaySort: $isDdaySort, sortingMethod: $sortingMethod)
-                    .padding(.bottom, 18)
-                ScrollView(.horizontal, showsIndicators: false) {
-                    HStack {
-                        TagsInFriendListView(bTags: bTags, relationshipTag: $relationshipTag, isshowingSheetForCreatingTag: $isshowingSheetForCreatingTag)
-                    } .padding(.leading, 25)
-                }
+                // MARK: - HeaderView
+                HeaderForFriendList(
+                    text: $text,
+                    sortingMethod: $sortingMethod,
+                    selectedViewMode: $selectedViewMode,
+                    isGridView: $isGridView
+                )
                 
-                Spacer()
+                TagsInFriendList(
+                    bTags: bTags,
+                    tagName: $tagName,
+                    tagColor: $tagColor
+                )
                 
-                
-                if isGridView {
-                    GridForFriendListView(bFriend: bFriend)
-                } else {
-                    ListForFriendListView(bFriend: bFriend)
-                }
+            
+                // MARK: - Grid Or ListView
+                FriendGridORListView(
+                    isGridView: $isGridView,
+                    text: $text,
+                    tagName: $tagName,
+                    sortingMethod: $sortingMethod,
+                    bFriend: bFriend
+                )
             }
+            .padding(.horizontal, 16)
+            .padding(.vertical)
             .background(Color.biRTH_mainColor)
+            
         }
+        
     }
 }
 
+
+#Preview {
+    FriendListView()
+}
